ansible-report utility now functional :)
Most of the code that formatted data for reporting is gone.  This
mostly collects the data that is needed and hands it off to functions in
utils for work.
This is very much oriented towards playbooks and tasks associated with
them.  Still to come are tasks that are not associated with a playbook
and other slicing/dicing of data.