Updates to manager.py
Drop sqlalchemy in favor of peewee.
Move lots of functionality out of model and into manager.
This includes:
  _filter_query()
  find_tasks()
  find_playbooks()
  get_task_stats()
  get_playbook_stats()
The callback plugin also leverages manager to do actual work of adding
to database:
  log_task()
  log_play()
  log_user()
A decorator is used with save() to catch errors or exceptions.  At this
point, the decorator is rudimentary but can be further enhanced.